requirements.txt
feedparser
requests
amadeus
python-dotenv
jinja2
resend
